requirements.txt
# Business Process Simulation Framework Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Scientific computing
scipy>=1.10.0

# Process mining and event log handling
pm4py>=2.7.0

# Configuration management
pyyaml>=6.0

# Simpy library for process simulation
simpy>=4.1.0

# Machine Learning (for duration prediction)
scikit-learn>=1.3.0

# Optimization solver (for scheduling)
ortools>=9.5.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0
